update filenames to class type names, move internal code into framework namespace
