+++ GeneralMods/BuildEndurance/Framework/ModConfig.cs
@@ -1,22 +1,22 @@
-﻿namespace Omegasis.BuildHealth
+namespace Omegasis.BuildEndurance.Framework
 {
     /// <summary>The mod settings and player data.</summary>
-    internal class Config
+    internal class ModConfig
     {
-        /// <summary>The XP points needed to reach the next level.</summary>
+        /// <summary>The XP points needed to reach the next endurance level.</summary>
         public double ExpToNextLevel { get; set; }
 
-        /// <summary>The player's current XP points.</summary>
+        /// <summary>The player's current endurance XP points.</summary>
         public double CurrentExp { get; set; }
 
-        /// <summary>The player's current level.</summary>
+        /// <summary>The player's current endurance level.</summary>
         public int CurrentLevel { get; set; }
 
-        /// <summary>The initial health bonus to apply regardless of the player's level, from the config file.</summary>
-        public int BaseHealthBonus { get; set; }
+        /// <summary>The initial stamina bonus to apply regardless of the player's endurance level.</summary>
+        public int BaseStaminaBonus { get; set; }
 
-        /// <summary>The health points to add to the player's base health due to their current level.</summary>
-        public int CurrentLevelHealthBonus { get; set; }
+        /// <summary>The stamina points to add to the player's base stamina due to their current endurance level.</summary>
+        public int CurrentLevelStaminaBonus { get; set; }
 
         /// <summary>The multiplier for the experience points to need to reach an endurance level relative to the previous one.</summary>
         public double ExpCurve { get; set; }
@@ -25,7 +25,7 @@
         public int MaxLevel { get; set; }
 
         /// <summary>The amount of stamina the player should gain for each endurance level.</summary>
-        public int HealthIncreasePerLevel { get; set; }
+        public int StaminaIncreasePerLevel { get; set; }
 
         /// <summary>The experience points to gain for using a tool.</summary>
         public int ExpForToolUse { get; set; }
@@ -35,6 +35,9 @@
 
         /// <summary>The experience points to gain for sleeping.</summary>
         public int ExpForSleeping { get; set; }
+
+        /// <summary>The experience points to gain for reaching a state of exhaustion for the day.</summary>
+        public int ExpForExhaustion { get; set; }
 
         /// <summary>The experience points to gain for collapsing for the day.</summary>
         public int ExpForCollapsing { get; set; }
