+++ GeneralMods/BuildHealth/BuildHealth.cs
@@ -2,47 +2,42 @@
 using System.IO;
 using System.Text;
 using Newtonsoft.Json;
+using Omegasis.BuildHealth.Framework;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
 
-namespace Omegasis.BuildEndurance
+namespace Omegasis.BuildHealth
 {
     /// <summary>The mod entry point.</summary>
-    public class BuildEndurance : Mod
+    public class BuildHealth : Mod
     {
         /*********
         ** Properties
         *********/
         /// <summary>The mod settings and player data.</summary>
-        private Config Config;
+        private ModConfig Config;
 
-        /// <summary>Whether the player has been exhausted today.</summary>
-        private bool WasExhausted;
-
-        /// <summary>Whether the player has collapsed today.</summary>
-        private bool WasCollapsed;
-
-        /// <summary>The XP points needed to reach the next endurance level.</summary>
+        /// <summary>The XP points needed to reach the next level.</summary>
         private double ExpToNextLevel = 20;
 
-        /// <summary>The player's current endurance XP points.</summary>
+        /// <summary>The player's current XP points.</summary>
         private double CurrentExp;
 
-        /// <summary>The player's current endurance level.</summary>
+        /// <summary>The player's current level.</summary>
         private int CurrentLevel;
 
-        /// <summary>The stamina points to add to the player's base stamina due to their current endurance level.</summary>
-        private int CurrentLevelStaminaBonus;
+        /// <summary>The health points to add to the player's base health due to their current level.</summary>
+        private int CurrentLevelHealthBonus;
 
-        /// <summary>The initial stamina bonus to apply regardless of the player's endurance level, from the config file.</summary>
-        private int BaseStaminaBonus;
+        /// <summary>The initial health bonus to apply regardless of the player's level, from the config file.</summary>
+        private int BaseHealthBonus;
 
         /// <summary>Whether to reset all changes by the mod to the default values (i.e. start over).</summary>
         private bool ClearModEffects;
 
-        /// <summary>The player's original stamina value, excluding mod effects.</summary>
-        private int OriginalStamina;
+        /// <summary>The player's original max health value, excluding mod effects.</summary>
+        private int OriginalMaxHealth;
 
         /// <summary>Whether the player recently gained XP for tool use.</summary>
         private bool HasRecentToolExp;
@@ -50,11 +45,14 @@ namespace Omegasis.BuildEndurance
         /// <summary>Whether the player was eating last time we checked.</summary>
         private bool WasEating;
 
+        /// <summary>The player's health last time we checked it.</summary>
+        private int LastHealth;
+
         /// <summary>Whether the player has loaded a save.</summary>
         private bool IsLoaded;
 
-        /// <summary>The player's stamina last time they slept.</summary>
-        private int NightlyStamina;
+        /// <summary>Whether the player has collapsed today.</summary>
+        private bool WasCollapsed;
 
 
         /*********
@@ -66,38 +64,37 @@ namespace Omegasis.BuildEndurance
         {
             GameEvents.UpdateTick += this.GameEvents_UpdateTick;
             GameEvents.OneSecondTick += this.GameEvents_OneSecondTick;
-            SaveEvents.AfterLoad += this.SaveEvents_AfterLoad;
-            TimeEvents.DayOfMonthChanged += this.TimeEvents_DayOfMonthChanged;
 
-            string configPath = Path.Combine(helper.DirectoryPath, "BuildEnduranceConfig.json");
+            TimeEvents.DayOfMonthChanged += this.TimeEvents_DayOfMonthChanged;
+            SaveEvents.AfterLoad += this.SaveEvents_AfterLoaded;
+
+            var configPath = Path.Combine(helper.DirectoryPath, "BuildHealthConfig.json");
             if (!File.Exists(configPath))
             {
-                this.Monitor.Log("Initial configuration file setup.");
-                this.Config = new Config
+                this.Monitor.Log("The config file for BuildHealth was not found, guess I'll create it...");
+                this.Config = new ModConfig
                 {
                     CurrentLevel = 0,
                     MaxLevel = 100,
-                    StaminaIncreasePerLevel = 1,
+                    HealthIncreasePerLevel = 1,
                     CurrentExp = 0,
                     ExpToNextLevel = 20,
                     ExpCurve = 1.15,
                     ExpForEating = 2,
                     ExpForSleeping = 10,
                     ExpForToolUse = 1,
-                    BaseStaminaBonus = 0,
-                    CurrentLevelStaminaBonus = 0,
-                    ExpForExhaustion = 25,
-                    ExpForCollapsing = 50
+                    BaseHealthBonus = 0,
+                    CurrentLevelHealthBonus = 0
                 };
                 File.WriteAllBytes(configPath, Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(this.Config)));
             }
             else
             {
-                this.Config = JsonConvert.DeserializeObject<Config>(Encoding.UTF8.GetString(File.ReadAllBytes(configPath)));
-                this.Monitor.Log("Found BuildEndurance config file.");
+                this.Config = JsonConvert.DeserializeObject<ModConfig>(Encoding.UTF8.GetString(File.ReadAllBytes(configPath)));
+                this.Monitor.Log("Found BuildHealth config file.");
             }
 
-            this.Monitor.Log("BuildEndurance Initialization Completed");
+            this.Monitor.Log("BuildHealth Initialization Completed");
         }
 
         /// <summary>The method invoked once per second during a game update.</summary>
@@ -131,13 +128,15 @@ namespace Omegasis.BuildEndurance
                 this.HasRecentToolExp = true;
             }
 
-            // give XP when exhausted
-            if (!this.WasExhausted && Game1.player.exhausted)
+            // give XP for taking damage
+            var player = Game1.player;
+            if (this.LastHealth > player.health)
             {
-                this.CurrentExp += this.Config.ExpForExhaustion;
-                this.WasExhausted = true;
-                this.Monitor.Log("The player is exhausted");
+                this.CurrentExp += this.LastHealth - player.health;
+                this.LastHealth = player.health;
             }
+            else if (this.LastHealth < player.health)
+                this.LastHealth = player.health;
 
             // give XP when player stays up too late or collapses
             if (!this.WasCollapsed && Game1.farmerShouldPassOut)
@@ -148,66 +147,37 @@ namespace Omegasis.BuildEndurance
             }
         }
 
-        /// <summary>The method invoked after the player loads a save.</summary>
-        /// <param name="sender">The event sender.</param>
-        /// <param name="e">The event data.</param>
-        public void SaveEvents_AfterLoad(object sender, EventArgs e)
-        {
-            // initialise
-            this.LoadConfig();
-            this.WriteConfig();
-            this.IsLoaded = true;
-
-            // grab initial stamina
-            var player = Game1.player;
-            if (this.OriginalStamina == 0)
-                this.OriginalStamina = player.MaxStamina;
-
-            // set nightly stamina
-            player.MaxStamina = this.NightlyStamina;
-            if (this.NightlyStamina == 0)
-                player.MaxStamina = this.BaseStaminaBonus + this.CurrentLevelStaminaBonus + this.OriginalStamina;
-
-            // reset if needed
-            if (this.ClearModEffects)
-                player.MaxStamina = this.OriginalStamina;
-
-            // save config
-            this.LoadConfig();
-            this.WriteConfig();
-        }
-
         /// <summary>The method invoked when <see cref="Game1.dayOfMonth"/> changes.</summary>
         /// <param name="sender">The event sender.</param>
         /// <param name="e">The event data.</param>
         public void TimeEvents_DayOfMonthChanged(object sender, EventArgs e)
         {
             // reset data
-            this.WasExhausted = false;
+            this.LastHealth = Game1.player.maxHealth;
             this.WasCollapsed = false;
             if (!this.IsLoaded)
                 return;
 
             // update settings
-            this.Monitor.Log(this.CurrentExp.ToString());
             this.UpdateClearSetting();
-            this.Monitor.Log(this.ClearModEffects.ToString());
 
-            var player = Game1.player;
+            var player = StardewValley.Game1.player;
             this.CurrentExp += this.Config.ExpForSleeping;
-            if (this.OriginalStamina == 0)
-                this.OriginalStamina = player.MaxStamina; //grab the initial stamina value
+            if (this.OriginalMaxHealth == 0)
+                this.OriginalMaxHealth = player.maxHealth; //grab the initial Health value
 
             if (this.ClearModEffects)
             {
                 this.LoadClearSettings();
-                player.MaxStamina = this.OriginalStamina;
+                //This will run when the character goes to sleep. It will increase their sleeping skill.
+                player.maxHealth = this.OriginalMaxHealth;
                 this.ExpToNextLevel = this.Config.ExpToNextLevel;
                 this.CurrentExp = this.Config.CurrentExp;
-                this.CurrentLevelStaminaBonus = 0;
-                this.OriginalStamina = player.MaxStamina;
-                this.BaseStaminaBonus = 0;
+                this.CurrentLevelHealthBonus = 0;
+                this.OriginalMaxHealth = player.maxHealth;
+                this.BaseHealthBonus = 0;
                 this.CurrentLevel = 0;
+                this.Monitor.Log("BuildHealth Reset!");
             }
 
             if (!this.ClearModEffects && this.CurrentLevel < this.Config.MaxLevel)
@@ -216,42 +186,70 @@ namespace Omegasis.BuildEndurance
                 {
                     this.CurrentLevel += 1;
                     this.CurrentExp = this.CurrentExp - this.ExpToNextLevel;
-                    this.ExpToNextLevel = (this.Config.ExpCurve * this.ExpToNextLevel);
-                    player.MaxStamina += this.Config.StaminaIncreasePerLevel;
-                    this.CurrentLevelStaminaBonus += this.Config.StaminaIncreasePerLevel;
-
+                    this.ExpToNextLevel =
+                        (this.Config.ExpCurve * this.ExpToNextLevel);
+                    player.maxHealth += this.Config.HealthIncreasePerLevel;
+                    this.CurrentLevelHealthBonus += this.Config.HealthIncreasePerLevel;
                 }
             }
             this.ClearModEffects = false;
-            this.NightlyStamina = Game1.player.maxStamina;
+
+            this.WriteConfig();
+        }
+
+        /// <summary>The method invoked after the player loads a save.</summary>
+        /// <param name="sender">The event sender.</param>
+        /// <param name="e">The event data.</param>
+        public void SaveEvents_AfterLoaded(object sender, EventArgs e)
+        {
+            // initialise
+            this.LoadConfig();
+            this.WriteConfig();
+            this.IsLoaded = true;
+
+            // grab initial health
+            var player = Game1.player;
+            if (this.OriginalMaxHealth == 0)
+                this.OriginalMaxHealth = player.maxHealth;
+
+            // set max health
+            player.maxHealth = this.BaseHealthBonus + this.CurrentLevelHealthBonus + this.OriginalMaxHealth;
+
+            // reset if needed
+            if (this.ClearModEffects)
+            {
+                player.maxHealth = this.OriginalMaxHealth;
+                this.Monitor.Log("BuildHealth Reset!");
+            }
+
+            // save config
+            this.LastHealth = Game1.player.maxHealth;
+            this.LoadConfig();
             this.WriteConfig();
         }
 
         /// <summary>Update the settings needed for <see cref="ClearModEffects"/> from the latest config file on disk.</summary>
         void LoadClearSettings()
         {
-            this.LoadConfig();
-            this.WriteConfig();
-
             if (!Directory.Exists(Path.Combine(Helper.DirectoryPath, "PlayerData")))
                 Directory.CreateDirectory(Path.Combine(Helper.DirectoryPath, "PlayerData"));
 
-            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildEndurance_data_{Game1.player.name}.txt");
+            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildHealth_data_{Game1.player.name}.txt");
             if (!File.Exists(path)) //if not data.json exists, initialize the data variables to the ModConfig data. I.E. starting out.
             {
-                Console.WriteLine("Clear Data Loaded could not find the correct file.");
+                this.Monitor.Log("The config file for BuildHealth was not found, guess I'll create it...");
 
                 this.ClearModEffects = false;
-                this.OriginalStamina = 0;
-                this.BaseStaminaBonus = 0;
+                this.OriginalMaxHealth = 0;
+                this.BaseHealthBonus = 0;
             }
             else
             {
-                //loads the BuildEndurance_data upon loading the mod
+                //loads the BuildHealth_data upon loading the mod
                 string[] text = File.ReadAllLines(path);
-                this.BaseStaminaBonus = Convert.ToInt32(text[9]);
+                this.BaseHealthBonus = Convert.ToInt32(text[9]);
                 this.ClearModEffects = Convert.ToBoolean(text[14]);
-                this.OriginalStamina = Convert.ToInt32(text[16]);
+                this.OriginalMaxHealth = Convert.ToInt32(text[16]);
             }
         }
 
@@ -261,14 +259,14 @@ namespace Omegasis.BuildEndurance
             if (!Directory.Exists(Path.Combine(Helper.DirectoryPath, "PlayerData")))
                 Directory.CreateDirectory(Path.Combine(Helper.DirectoryPath, "PlayerData"));
 
-            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildEndurance_data_{Game1.player.name}.txt");
+            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildHealth_data_{Game1.player.name}.txt");
             if (!File.Exists(path)) //if not data.json exists, initialize the data variables to the ModConfig data. I.E. starting out.
             {
-                Console.WriteLine("Clear Data Loaded could not find the correct file.");
+                this.Monitor.Log("The config file for BuildHealth was not found, guess I'll create it...");
 
                 this.ClearModEffects = false;
-                this.OriginalStamina = 0;
-                this.BaseStaminaBonus = 0;
+                this.OriginalMaxHealth = 0;
+                this.BaseHealthBonus = 0;
             }
             else
             {
@@ -278,51 +276,52 @@ namespace Omegasis.BuildEndurance
         }
 
         /// <summary>Load the configuration settings.</summary>
-        void LoadConfig()
+        private void LoadConfig()
         {
             if (!Directory.Exists(Path.Combine(Helper.DirectoryPath, "PlayerData")))
                 Directory.CreateDirectory(Path.Combine(Helper.DirectoryPath, "PlayerData"));
 
-            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildEndurance_data_{Game1.player.name}.txt");
+            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildHealth_data_{Game1.player.name}.txr");
             if (!File.Exists(path)) //if not data.json exists, initialize the data variables to the ModConfig data. I.E. starting out.
             {
+                this.Monitor.Log("The config file for BuildHealth was not found, guess I'll create it...");
                 this.ExpToNextLevel = this.Config.ExpToNextLevel;
                 this.CurrentExp = this.Config.CurrentExp;
                 this.CurrentLevel = this.Config.CurrentLevel;
-                this.BaseStaminaBonus = this.Config.BaseStaminaBonus;
-                this.CurrentLevelStaminaBonus = this.Config.CurrentLevelStaminaBonus;
+                this.BaseHealthBonus = this.Config.BaseHealthBonus;
+                this.CurrentLevelHealthBonus = this.Config.CurrentLevelHealthBonus;
                 this.ClearModEffects = false;
-                this.OriginalStamina = 0;
+                this.OriginalMaxHealth = 0;
             }
             else
             {
-                // loads the BuildEndurance_data upon loading the mod
+                //loads the BuildHealth_data upon loading the mod
                 string[] text = File.ReadAllLines(path);
                 this.CurrentLevel = Convert.ToInt32(text[3]);
-                this.ExpToNextLevel = Convert.ToDouble(text[7]);  //these array locations refer to the lines in BuildEndurance_data.json
+                this.ExpToNextLevel = Convert.ToDouble(text[7]);  //these array locations refer to the lines in BuildHealth_data.json
                 this.CurrentExp = Convert.ToDouble(text[5]);
-                this.BaseStaminaBonus = Convert.ToInt32(text[9]);
-                this.CurrentLevelStaminaBonus = Convert.ToInt32(text[11]);
+                this.BaseHealthBonus = Convert.ToInt32(text[9]);
+                this.CurrentLevelHealthBonus = Convert.ToInt32(text[11]);
                 this.ClearModEffects = Convert.ToBoolean(text[14]);
-                this.OriginalStamina = Convert.ToInt32(text[16]);
-                this.NightlyStamina = Convert.ToInt32(text[18]); //this should grab the nightly stamina values
+                this.OriginalMaxHealth = Convert.ToInt32(text[16]);
             }
         }
 
         /// <summary>Save the configuration settings.</summary>
-        void WriteConfig()
+        private void WriteConfig()
         {
-            // saves the BuildEndurance_data at the end of a new day
             if (!Directory.Exists(Path.Combine(Helper.DirectoryPath, "PlayerData")))
                 Directory.CreateDirectory(Path.Combine(Helper.DirectoryPath, "PlayerData"));
 
-            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildEndurance_data_{Game1.player.name}.txt");
+            string path = Path.Combine(Helper.DirectoryPath, "PlayerData", $"BuildHealth_data_{Game1.player.name}.txt");
             string[] text = new string[20];
             if (!File.Exists(path))
             {
-                Console.WriteLine("The data file for BuildEndurance was not found, guess I'll create it when you sleep.");
+                this.Monitor.Log("The data file for BuildHealth was not found, guess I'll create it when you sleep.");
 
-                text[0] = "Player: Build Endurance Data. Modification can cause errors. Edit at your own risk.";
+                //write out the info to a text file at the end of a day. This will run if it doesnt exist.
+
+                text[0] = "Player: Build Health Data. Modification can cause errors. Edit at your own risk.";
                 text[1] = "====================================================================================";
 
                 text[2] = "Player Current Level:";
@@ -334,26 +333,24 @@ namespace Omegasis.BuildEndurance
                 text[6] = "Xp to next Level:";
                 text[7] = this.ExpToNextLevel.ToString();
 
-                text[8] = "Initial Stam Bonus:";
-                text[9] = this.BaseStaminaBonus.ToString();
+                text[8] = "Initial Health Bonus:";
+                text[9] = this.BaseHealthBonus.ToString();
 
-                text[10] = "Additional Stam Bonus:";
-                text[11] = this.CurrentLevelStaminaBonus.ToString();
+                text[10] = "Additional Health Bonus:";
+                text[11] = this.CurrentLevelHealthBonus.ToString();
 
                 text[12] = "=======================================================================================";
                 text[13] = "RESET ALL MOD EFFECTS? This will effective start you back at square 1. Also good if you want to remove this mod.";
                 text[14] = this.ClearModEffects.ToString();
-                text[15] = "OLD STAMINA AMOUNT: This is the initial value of the Player's Stamina before this mod took over.";
-                text[16] = this.OriginalStamina.ToString();
-
-                text[17] = "Nightly Stamina Value: This is the value of the player's stamina that was saved when the player slept.";
-                text[18] = this.NightlyStamina.ToString(); //this should save the player's stamina upon sleeping.
+                text[15] = "OLD Health AMOUNT: This is the initial value of the Player's Health before this mod took over.";
+                text[16] = this.OriginalMaxHealth.ToString();
 
                 File.WriteAllLines(path, text);
             }
             else
             {
-                text[0] = "Player: Build Endurance Data. Modification can cause errors. Edit at your own risk.";
+                //write out the info to a text file at the end of a day.
+                text[0] = "Player: Build Health Data. Modification can cause errors. Edit at your own risk.";
                 text[1] = "====================================================================================";
 
                 text[2] = "Player Current Level:";
@@ -365,20 +362,17 @@ namespace Omegasis.BuildEndurance
                 text[6] = "Xp to next Level:";
                 text[7] = this.ExpToNextLevel.ToString();
 
-                text[8] = "Initial Stam Bonus:";
-                text[9] = this.BaseStaminaBonus.ToString();
+                text[8] = "Initial Health Bonus:";
+                text[9] = this.BaseHealthBonus.ToString();
 
-                text[10] = "Additional Stam Bonus:";
-                text[11] = this.CurrentLevelStaminaBonus.ToString();
+                text[10] = "Additional Health Bonus:";
+                text[11] = this.CurrentLevelHealthBonus.ToString();
 
                 text[12] = "=======================================================================================";
                 text[13] = "RESET ALL MOD EFFECTS? This will effective start you back at square 1. Also good if you want to remove this mod.";
                 text[14] = this.ClearModEffects.ToString();
-                text[15] = "OLD STAMINA AMOUNT: This is the initial value of the Player's Stamina before this mod took over.";
-                text[16] = this.OriginalStamina.ToString();
-
-                text[17] = "Nightly Stamina Value: This is the value of the player's stamina that was saved when the player slept.";
-                text[18] = this.NightlyStamina.ToString();
+                text[15] = "OLD Health AMOUNT: This is the initial value of the Player's Health before this mod took over.";
+                text[16] = this.OriginalMaxHealth.ToString();
 
                 File.WriteAllLines(path, text);
             }
