+++ GeneralMods/BuyBackCollectables/BuyBackCollectables.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using Omegasis.BuyBackCollectables.Framework;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
