+++ GeneralMods/DailyQuestAnywhere/DailyQuestAnywhere.cs
